refactor(ast): reorganize AST module into subdirectory structure
--ai

Reorganize the AST module from a single mod.rs file into a hierarchical
subdirectory structure to improve maintainability and extensibility.

Changes:
- Move Node trait and Program struct into ast/mod.rs
- Extract Statement enum and LetStatement into ast/statement/ subdirectory
- Extract Expression enum and Identifier into ast/expression/ subdirectory
- Update import paths in lib.rs to use new module structure

This structure makes it easier to add new statement and expression types
by creating new files in their respective subdirectories, reducing merge
conflicts and improving code organization.

Structure:
- ast/mod.rs: Node trait, Program struct
- ast/statement/: Statement enum, LetStatement
- ast/expression/: Expression enum, Identifier
